servicerequirements.txt
amqp==5.1.1
aniso8601==9.0.1
billiard==3.6.4.0
CacheControl==0.12.11
cachetools==5.2.0
celery==5.2.7
certifi==2022.9.24
charset-normalizer==2.1.1
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
firebase-admin==6.0.0
Flask==2.0.1
Flask-RESTful==0.3.9
google-api-core==2.10.2
google-api-python-client==2.64.0
google-auth==2.12.0
google-auth-httplib2==0.1.0
google-cloud-core==2.3.2
google-cloud-firestore==2.7.1
google-cloud-storage==2.5.0
google-crc32c==1.5.0
google-resumable-media==2.4.0
googleapis-common-protos==1.56.4
grpcio==1.49.1
grpcio-status==1.49.1
httplib2==0.20.4
idna==3.4
itsdangerous==2.0.1
Jinja2==3.0.1
kombu==5.2.4
MarkupSafe==2.0.1
msgpack==1.0.4
neo4j==4.4.0
peewee==3.14.4
prompt-toolkit==3.0.31
proto-plus==1.22.1
protobuf==4.21.7
pyasn1==0.4.8
pyasn1-modules==0.2.8
PyJWT==2.5.0
pyparsing==3.0.9
pytz==2022.1
requests==2.28.1
rsa==4.9
six==1.16.0
uritemplate==4.1.1
urllib3==1.26.12
vine==5.0.0
wcwidth==0.2.5
Werkzeug==2.0.1
python-dotenv==1.0.1